feat: Automated Badge Achievement System
