+++ src/types/index.ts
@@ -57,6 +57,8 @@ export interface OrgStats {
   issuesCount?: number;
   commitsCount?: number;
   reviewsCount?: number;
+}
+
 export interface GistItem {
   id: string;
   description: string | null;
